src/mc68/pmon: Define memory locations externally to this file
Including files can (and must) now set several constants to determine
where the three RAM areas (low and high general storage and the input
buffer) and the code are placed.

Both the unit test and tmc68 versions now define more sensible
locations for all of these. (This means that this code will now
overwrite important locations on the JR-200, though; we need to create
a proper version for that platform.)

A bit of trickery with the `pushv` directive gives us clearer error
messages and aborts the assembly when the required constants are not
defined.