cmtconv.data: BlockHeader and Block classes extracted from cmtconv.jr200
This is the first step in starting to clarify the layering of the code
(files/blocks, byte streams, bit streams and audio data). We choose a more
generic name for the module with the idea that when we add the next
platform the current module will give us the API and we'll be able to see
what parts need to be separated out into platform-specific modules.

We extract this small bit first with the aim of adding unit tests and
bringing all the block handling together. (Right now it appears that the
FileHeader is a sort of block, but is completely separate from the
BlockHeader, and various other stand-alone functions such as
checksums_valid are actually block-related.)

The one debug() call in this code has been replaced with a call to
logging.debug() which currently produces no output. In the future we should
add a command line option to set the level; in the meantime it can be
turned on by temporarily adding the following to an appropriate place in
the code:

    logging.getLogger().setLevel(logging.DEBUG)