cmtconv.data: Blocks no longer take a checksum
The Block no longer holds the (possibly incorrect) checksum from a read;
instead we compare the read checksum to the Block's calculated checksum at
the time we do the read and fail immediately (with a RuntimeError) if it
doesn't match. (Later we can change to emitting a warning and carrying on
or whatever, if we like.)

This eliminates checksums_valid(); it's no longer necessary to remember to
call it after reading blocks because the Block objects were already checked
when they were created.

- Block.calc_checksum() is now checksum()
- Block.__repr__() replaces Block.__str__ with a more compact
  representation, but this will go away anyway when we merge Block and
  BlockHeader.
- BlockHeader.calc_checksum() is gone; it never made much sense since the
  checksum can be calculated only over the whole block, header and data.