src/mc68/pmon: Separate tests to confirm commands correctly use pending value
The way the pending value is used is common to all commands: use a
command-specific current address if no pending value is supplied, otherwise
copy the pending value to that current address and use it. So we extract a
separate test for this parametrized across the various commands and their
current location variables, allowing us to simplify and use fewer tests for
the commands themselves.

This also means that the function that sets up the command argument input
no longer needs to deal with the pending value, so we remove the
pending value handling from that and rename it to set_rdlinebuf().

We also tweak set_rdlinebuf() buf now to use different offsets from 0-7
into the rdlinebuf, determined more or less randomly from the contents of
the input string. This helps ensure that commands aren't accidentally using
the start of the buffer instead of the pointer to the current location in
the buffer.