src/mc68/pmon: Add width option to `x`; length opt now needs `L` prefix
The options after `x` are now case-insensitive single characters `L` for
length and `W` for width followed by a hexbyte (1 or 2 chars) to
(persistently) set that option.

The code for each is currently duplicated. Before addressing this, we need
to try out a few more options, such as switches for ASCII etc. display, and
ideally options for other commands too. Then we can figure out issues with
error handling and so on, settle on a convenient syntax, and then extract
routines to help deal with this. (It's not clear right now what these
routines should look like. E.g., should they be table-driven?)

With the current parsing style, options, like commands, may not use letters
`A`-`F` (or numbers). So far this hasn't been a problem.