src/mc68/pmon: Register display
We needed to drop the idea of using an `=` sign between the register name
and value in order to get the output to fit within 32 chars, for systems
with short screen lines. This feels like it may be harder to read on
systems that don't support lower-case, but I'm not sure how to deal with
that. Perhaps print the register name in inverse?

Also, 73 bytes feels like a lot just to display the registers. Perhaps it
can be be made tinier by tweaking the storage order or by sharing some
output routines with examine or something like that.