src/mc68/pmon regscmd: Functionality to set saved registers
We change the command from `s` to `r` because the latter is just much more
instinctive and easier to remember. We deal with the loss of R)ead by
changing read, write and verify all to be under the T)ransfer command,
which should be fine because that command will have a ton of other options
anyway.

There's still a design problem with this, as showed up with some
frustrating debuggin during unit testing: since the args are always
adjacent, it's easy to type `a3b4` meaning to set A=3 and B=4, but that's
interpreted as A=3B 4= followed by a parse error. Currently, we're thinking
of allowing optional commas in to separate arguments to help mitigate this.

The code is a bit sloppy and has a good deal of duplication, and the tests
are showing issues similar to (but not the same as) the examine tests. But
this is going on master anyway because it's unlikely to have broken
anything working to this point and we want feedback on this sooner rather
than later.