testmc.*.machine: Add _getsp(), pushretaddr(), getretaddr()
Machine._getsp() is a machine-independent way of getting the current
value of the stack pointer, and also may be implemented faster than
using a Registers object, just as with _getpc(). (Doing it quickly is
important to speed future versions of Machine.call().)

Machine.pushretaddr() pushes an RTS return address on to the stack,
and getretaddr() returns the current head of the stack interpreted as
a return address for RTS.

These are machine-specific for not only word order and interpretation
of the stack pointer (e.g., whether it points at or below the head of
the stack) but also for the address itself. For example, the 6502
return addresses when stored on the stack are 1 less than the actual
return address.

These all will be used by a new version of call() that tracks stack
usage directly, rather than via the proxy of watching JSR and RTS
instructions. (Some programs may stack return addreses via push rather
than JSR, or remove return addresses with pop, which confuses the
current call() implementation.)