testmc.m6502.machine call(): Better tests and move down to each CPU
The new implementation of testmc.generic.machine.call() will rely on
correct implementation of some machine-specific stack support, so it's best
to be running the full set of tests on every Machine implementation, and
abandon the attempt to have generic test versions of JSR, RTS, etc.
(especially since we'd need to add push/pull/whatever to test the new
version of call).

So we remove all the existing minimal call() tests on testmc.{m6502,mc6800}
and replace them with a better set of full tests based on the original
testmc.generic call() tests. In particular, the deep call test is improved
to cover all subroutine call instructions (JSR and BSR on MC6800), include
a regular branch in the middle, and be a bit recursive. (This makes it
slightly hard to understand, but I'm not sure how to make it easier without
being more verbose.)