testmc.*.machine tests: Make R (Registers) a global instead of a fixture
This means we no longer need pass it as a parameter to every test
function that uses it.

We do not do this for testmc.generic.machine because that's a bit
special; it requires a newly defined Registers class for several of
the tests because it modifies that class and we don't want those
changes leaking into other tests.