testmc.*.machine: Re-implement call()
This new version no longer counts JSRs and RTSs, but instead places a
sentinel return address on the stack and stops execution when either:
1. We reach an RTS with the stack pointing to the location we placed
   the sentinel address (regardless of the contents of the stack at
   that location), or
2. We reach an RTS with the stack containing the sentinel address
   (regardless of the location to which the stack pointer points)

If both conditions above are true we return successfully. If only one
is true there's been some stack manipulation that may indicate an
error and we raise an UnexpectedCallStack exception. (This can be
caught in the unlikely case where we expect the code under test to RTS
with the stack in a different condition than when it was entered.)

The new situations this handles are:
- test_call_pull(): CUT executing a JSR with no corresponding RTS.
- test_call_push(): CUT executing an RTS with no corresponding JSR.
- test_call_changedstack(): CUT executes RTS to the sentinel return
  address call() placed on the stack, but moved to a different
  location on the stack.
- test_call_changedretaddr(): CUT clobbers the return address call()
  placed on the stack and then executes RTS.

The Machine._JSR_opcodes() method is now no longer needed (since we no
longer look at JSRs) and thus is removed.

One test under src/ depended on the old stack behaviour of call() and
needed to be tweaked to expect the new behaviour.