cmtconv.jr200: Change output to fix edge detection on roundtripping
Edge detection when dumping a WAV file was broken because the last
edge was not being detected when going to silence. Work around this by
generating another edge at the end of the output.

Now that cmt-test's test_wavtodump is successfully producing output,
update the test to compare that with expected output.