testmc.generic stepto(): Return the number of steps executed
Callers with their own `maxsteps` value that call stepto() multiple times
need to track how many steps are executed across across all the multiple
calls. This allows them to do that.

Currently call()'s `maxsteps` is broken because it sets that full limit for
each individual call to stepto(), so it may even go on forever if no
individual stepto()'d fragment exceeds `maxsteps`. We don't fix this now
because we're about to rewrite call(), but we need to add a test for this
later.