src/mc68/pmon: Replace runmon() with call()
Now that call() properly handles code that manipulates the stack, we can
just call(S.init) rather than setting up the PC and using stepto() with an
exit point. This is also more reliable as it will work with commands that
use a different RTS instruction to exit.

This does drop the `maxsteps=20000` that every call was previously using,
but it's not an issue with all tests working, and we can always add it back
in later at indivdiual call sites, if necessary.