src/a2/mondump: Print 8 hex bytes and one text byte
This is old code (from about eight months ago) that was sitting on a
branch that was never pushed up. It's had a quick check by hand in
linapple to see that it works as far as it goes, but the unit test
needs working COUT, PRBYTE etc to test this and develop it further.