src/mc68/pmon prperr: Print error message with parse error information
This is currently only being printed when we fail to parse a command,
but it can easily be used for bad options as well, since in that case
too we want to abort further command line processing by restarting the
command loop.

Also slightly improve the debug output of set_rdlinebuf() in the
tests, to give us both the start of the rdline buffer and the offset
at which it placed the input.