src/mc68/pmon: eXamine terminated by whitespace; parse error on bad arg
We used to allow two things in examine command argument parsing that turn
out to be a bad idea:
- Allow a bare hexbyte to be the L)ines argument value.
- Stop parsing and assume success when we get a bad argument parse.

The first means that if someone misremembers W)idth as C)olumns or
something like that it sets L)ines and then things get confusing from
there. The second means that if someone intended something they typed to be
an argument for examine, but made an error, it instead may successfully
execute as a subsequent command, which also introduces confusion.

So we tighten things up a bit to reduce ambiguity about what apparent
errors were really supposed to mean by declaring:
- All commands that take arguments must be terminated with whitespace.
  ("Whitespace" includes end of line.)
- All commands that take multiple arguments do not also allow a bare
  hexbyte or hexword as an argument.
- Any command that has a parse error on its arguments now aborts processing
  of rest of the command line. (This is obviously necessary for safety.)

With this design in place, we now add parse error messages to examine
command arguments in a fairly trivial way: just call the same prperr_up
routine that the main loop falls into on a parse error. (Remember, this
removes the continuation that would continue parsing the command line and
then prints an error indicating the offending letter and its offset in the
rdline buffer.)

The testing has gotten awkward, and it's now clear that we should delink
the testing of examine's argument parsing and its output, per the comments
in the tests. We leave this cleanup for a separate commit.