src/mc68/pmon: New \xC0 (print `@`) command for testing
This removes the confusion in tests that was caused by `^` as the prompt
and `^` as the echo of a print newline command. Also, using a character
with the high bit set ($C0 = '@'∙$80) ensures that it won't collide with or
use up characters needed by any real commands we add later.