src/mc68/continuation: framework for continuations with demo and explanation
This is a proof of concept that shows that the idea works. The subroutine
and macros should be pulled out to a library file at some point.

The explanations may not be as good as they could be, and while the example
demonstrates _how_ the system works, another (more complex) example is
needed to show _why_ this framework/technique is helpful. (There is an
explanation of how pmon uses it, but the pmon code itself may or may not be
clear enough to be a good example.)