src/*/std.a*: Add {NOEXPAND} to `ds` macros
There's no need to display a second line of `rmb` in the listing when using
the `ds` macro: the address at the left on the `(MACRO)` line shows the
first address being reserved.

The `zds` macro already had {NOEXPAND}, which is even more useful there
because it expands to seven lines of code. However, lack of expansion is a
problem there, because without expansion the address column in the listing
shows only the current assembly address where `zds` is called, not the
address actually generated for the symbol. We'll have to see if we can find
a way to mitigate this.