src/m65/std: Explanation of ZDS macro expansion dilemma
We can't see the location assigned at definition point, and several tries
to work around this without listing all seven lines of the macro did not
work out. So give a workaround for the moment.

A probable bug in AS 1.42 bld 172 was discovered: LISTON is supposed to be
a dynamic variable, but using LISTING ON or RESTORE after a LISTING OFF
raises an error that LISTON can't be redefined. This should be reported to
the mailing list.