src/asl/simple: Clean up negoff() test
There's no need to run code that uses negoff() and stores the value; we can
just define a symbol using negoff() on a value and read the symbol value
that was assembled.

negoff(), which lets us count upto an end point when indexing rather than
downto one, should probably be part of m65/std.a65.