src/mc68/pmon: startline/inline to better format output
This avoids unnecessary newlines when we are processing multiple commands
on a line, some of which want to do output on their own lines and some of
which have short output that should go at the end of the current line.

The `stat` byte gains a `stat_begline` bit which is set when we know we're
at the beginning of a line. (Thus, it's clear when `stat` is initialized to
$00, which is correct because at startup from an unknown state we prefer an
unnecessary newline to starting in the middle of a line.)

New subroutines are:
- nextline: Always print a newline and set begline.
- startline: Print a newline unless begline is set, and set begline.
- inline: Clear begline to indicate that output without a trailing
  newline has been or will be generated.

This isn't totally automatic since routines that generate output
without a trailing newline need to remember to call `inline`, but
that's less overhead and a lot less hassle than wrapping all the BIOS
output routines we're likely to call.

The tests are currently a bit awkward about how the test that the
expected status bits are being changed (and the others not changed).
We need to improve this.