cmtconv.data.BlockHeader: Rename `length` to `datalen`
This makes it clear that the length in question is not the length of the
entire block but only of the data payload. (The comments describing the
block header have also been updated.)

We store this internally as self._datalen for the moment and use `@property
def datalen(self)` the accessor; after we merge BlockHeader with Block this
can simply return len(data) rather than duplicating the length value with
the possibility of being inconsistent.