BROKEN cmtconv: Merge BlockHeader into Block
cmtconv.jr200 has not yet been changed to use the new API.

This now gives a new API to the clients that relieves them from having to
know any block header internals, beyond the block header length. They now
hand the header bytes to Block.from_header() which returns a block with
empty data and the datalen from the binary header. The client then reads
that many bytes plus the checksum byte and passes those two to setdata() on
the block, which now handles checksum checks as well.