WIP cmtconv.jr200: Convert to use new cmtconf.data API
There's still a good deal of cruft code (and some commented-out code)
dealing with the details that Block now deals with. This is mostly marked
with XXX and needs to be cleaned up.

The API was tweaked slightly because we now see it's better for
TailBlock.data to return b'' instead of None.