cmtconv.data.Block: Allow any sequence type for block data params
There's really no good reason we shouldn't make it easy for others to
provide sequences other than `bytes` for the data for a block. (Pushing
type suggestions upwards really isn't such a great idea in languages like
Python.) So just pass whatever we get to bytes() which will produce an
appropriate exception if there are any values too small or large for a
byte.