WIP cmtconv.{block,bytestream}: file and block data bytestream functions
XXX what remains to do on this?

cmtconv.bytestream is a module of generic functions that read and generate
streams of bytes representing either tape file contents or a series of
blocks on tape. It's now the interface to the platform-specific Block
classes, which live under cmtconv.block.

The JR-200 block functions that read and generate these bytestreams move to
cmtconv.block.jr200, where they can be auto-loaded by cmtconv.bytestream.