cmtconv.data: Rework public API; replace .data() with .filedata
In further development, some confusion arose between the contents of the
data portion of a block on tape and the file data contained in blocks,
because for some blocks the data portion contains metadata, not file data.
We resolve this by removing direct read access to the data portion from the
public API (removing Block.data() and Block.datalen()) and instead
providing a Block.filedata property  to read the file data contained in the
block, if any. Blocks containing only metadata (or no data at all) will
return a zero-length `bytes` when Block.filedata is read. (Block.setdata()
remains; this is used when reading the raw bytestream from tape audio or an
equivalant block bytestream.)

Additional changes include:
- Renaming "tail" to "eof" because this is more clear about what the block
  means (the end of a file on a tape).
- Changing all the public read accessors to be properties. You now read
  Block.checksum and Block.is_end rather than calling methods with those
  names.
- Renaming `make()` to `make_block()` for slightly improved clarity.
- Adding reasonable default values for FileHeader.make_block().

Direct access to the data portion of the block remains as the "private"
property Block._data; this is intended for internal use only, and is not
part of the public API.

This is also now better documented in the docstrings.