cmtconv.logging: Output for different verbosity levels
This re-uses the Python standard library `logger` package for our purposes,
with our parts configured to be separate from standard Python logging. See
the module docstring for more details.

It's not very well tested because it's hard to unit test, but is not
critical to operation. We can fix infelicities as we notice them during
development and debugging of tape formats.