bin/cmtconv: Start on single program to do all conversions
This currently can read a 'cas' format file (containing the bytes
representing the data blocks on tape) and write it out as a 'bin' file
(containing the contents of that tape file).

It's in very rough shape, but the general concept that seems to be emerging
is that one reads from any format to produce a sequence of Block objects,
and then writes information from those out to any format.

This introduces cmtconv.formats to manage the format names and their
corresponding read/write functions.