BROKEN hw/mempcb/: initial component placement; up-side down bus conn.
The RC6502 bus connector is upside-down compared to normal boards, with the
pins going the wrong direction when looking at the component side as
compared with all other RC6502 boards. This must be fixed because otherwise
it's almost certain to lead to someone inserting a board into the backplane
backwards.