hw/mempcb: Move RC6502 header to grid, and shift board under it
The RC6502 bus header's pins not being centred on the 1.27 mm grid was
causing all sorts of annoyance with the trace routing near it. So we shift
it slightly to the right to get it on that grid, consistent with all the
other components, and then move the board outline slightly to keep the
header entirely on the board.

(Actually, the component outline hangs just slightly over the right edge of
the board, but that shouldn't be a problem. If we shift it a smidgen
further so that instead the spare/IO-select pin 39 hangs over we should
make sure to check that the 50-pin microcomputer bus, the outside row of
which is directly above pin 39, is still ok.)

The traces for the RC6502 bus header are now borked, and need to be fixed
by hand.