WIP hw/mempcb: Move power connectors; ARGH! DIP-32 is not aligned to grid!
Moving the power connectors up to the top edge of the board lets us hook
them with under-board traces to the DIP-32 Vcc. (Back in the other corner
there was no way to route power to them without topside traces, so clearly
they don't want to go there.)

It's kind of looking like the Vcc stuff, including the pull-ups, wants
to be concentrated at the centre of the board. We could move the right
logic socket over to allow this to happen.