hw/mempcb: Moved right logic socket to far right of board
This appears to have worked out well. The pull-ups and extra Vcc/GND pins
are in now in the centre, between the logic sockets, but not in their final
places because as it turns out I want to change around how the Vcc/GND pins
are set up, splitting them on to two separate headers. (We probably should
have done that in the first place, since a pair of headers can be placed
adjacent on the board to make what becomes one longer header.