hw/mempcb: New GND/Vcc WW pins: 4×GND, 2×Vcc, side by side w/buspwr jumper
The BUS_VCC jumper is now directly above the RC6502 bus Vcc pins, and the
two wire-wrap Vcc pins are directly above that. Directly to its right,
above the RC6502 bus GND, we have four wire-wrap pins for GND. (These
latter have traces on the bottom just to make this easier to recognise when
looking at the backside; they still should be connected to the top ground
fill when we get to that.)

These and the pull-up wire-wrap headers are oriented vertically to make
routing of wire-wrap easier.

All of these, including the pull-ups, are now routed.