hw/mempcb: Reroute D0-D2 to remove a crossing trace; only GNDs left
By using a narrower trace spacing we have plenty of room to run all
the data lines to the left of the left logic chip, allowing the Vcc
pin of that chip now to be entirely a bottom-side trace.

Also:
- Some general cleanup of the data lines.
- Align the right-side logic chip with the grid.
- Make the A10 via cross only two lines instead of three.

The only nodes left in the unconnected list seem to be GND pins.