hw/mempcb-cache/mempcb-cache.lib: Remove library prefixes from names
In the previous verison of this file, created by KiCAD 5.1, each item in
the cache has a name prefixed by the library name, rather than the name
directly from the library. E.g.:

    R                   → Device_R
    Conn_02x25_Odd_Even → Connector_Generic_Conn_02x25_Odd_Even

KeyCAD 5.0 doesn't seem to understand this and rescues all of these
symbols. However, on this commit, with the prefix removed, KiCAD 5.0 loads
this schematic with no errors.

This commit will let us test to see what KiCAD 5.1 does with this.