WIP hw/mempcb: Power trace under bus now 0.22 mm from boad edge
Pin 1 (A15) for the RC6502 bus header is now at absolute 150,200, for
easier counting. The new 20 mil board-edge Vcc trace is directly below it,
just clearing the pads and confirmed ok by the DRC. The board edges have
been moved to give .22 mm of space below that trace; this is not checked in
5.0 but may be checked in 5.1.

Ignore previous notes about grid origin; that's set by the user to whatever
he finds convenient at the moment. (See sedoc for more details.)

Still to do:
- Route Vcc around other edges of board.
- Move parts to new grid centred on RC6502 header pin 1.
- Reconnect all the broken traces.