hw/mempcb: Widen board edge to 0.4 mm; tweak corners
According to [1], fabs cut along the _center_ of the board edge line, so
you can make it as wide as you want. Thus a cheap and easy way of ensuring
you maintain margin at the edge of the board is to widen the board line to
twice the margin you need, because KiCad will not let traces etc. _enter_
the line.

I can't check this on 5.0 because the DRC doesn't check for traces
exceeding the board boundary, but I believe that 5.1 does check for this,
so we can test it there.

[1]: https://forum.kicad.info/t/margin-or-edge-cuts/10341/2