hw/mempcb: Rename most references and remove some from silkscreen
We rename the ICs to U1, U2 and U3 so that it doesn't look like some are
missing. We resolve the previous collisions with the wire-wrap headers by
renaming them all to WW_*.1 and WW_*.2 where * is the part with which
they're associated. The reference alignment and locations are also cleaned
up a bit.

Update the PCB from this, make invisible the skilk screen references for
the WW_* headers (since their associated part is easily clear from
proximity) and for the others move the silk screen to locations that are
more clear.