tool/dos33fs/Setup: Fix for early Jan. directory structure changes
The directories were restructured in early January this year to put
things under a new layer of subdirectories, which requires only that
we fix our bin/ links.

Note that some versions of dos33fs from around 2021-01-05 will not
build; the master branch in that repo was broken for a while during
the restructuring.