src/mc68/pmon qcmdopts: extract subroutine to parse command options+values
The new routine, `qcmdopts`, takes takes a table describing the option
letters, the type for each option, and the address at which the option's
value should be stored. Currently byte and word values are parsed; these
have types bytes with more than one bit set so that we can later use single
bits as indicators of both flag type and which bit within the value address
should be set/reset.

If the command options table contains a bad type, we currently just execute
an illegal instruction which in the CPU simulator causes an exception. That
works fine at catching the problem when the option is unit-tested, but
would produce very mysterious failures if run on a real system, so perhaps
we should find something better to do.

We include the correct options table for the regs command, but we're
delaying making that command actually use qcmdopts in order to make this
commit less confusing.

This introduces a new test helper function, prperr_abort_check(), which
makes it easy to assert that parsing fails at the correct place (i.e., that
prperr_abort is called at the appropriate point) when we expect a parse
error. Later we should clean up other tests to use this.