src/mc68/pmon: Rework prperr code to handle arbitrary stack depth on abort
The old prperr_up routine assumed we were at a specific stack depth when
doing stack cleanup after being called to abort further parsing of the
command line, which meant that it was very tricky to call.

Instead, we now simply save the stack pointer before calling a command so
that the routine, now renamed to prperr_abort, properly drops any arbitrary
stack state, making it callable from any stack depth in a command. This is
basically setting the catch point for command-generated exceptions. This
makes it easier to call and much more reliable.

This has no effect on commands that do not abort, allowing them to do
things like move the stack and restart the monitor, or even continue
command-line parsing if they copy the parser continuation return address.