src/mc68/pmon regcmd: Use qcmdopts
This also fixes a minor bug in regcmd's old options parsing code that did
not give the right failure position when a register name was given with an
EOL rather than a value following it. (It was showing the parse failure at
the register name, even though that was parsed fine, rather than the
end-of-line where the value was supposed to be.)

test_regscmd_set() should be at least cleaned up, or possibly eliminated,
since the qcmdopts and prperr tests are covering most of what it does.