src/mc68/pmon examine: Use qcmdopts
We also remove the upper-case aliases of the options (`L` and `W`) since
there doesn't seem to be a lot of point to them.

test_examine_parseargs() is now doing far too much work, duplicating much
of the coverage already provided by test_qcmdopts(). Possibly we could live
without the test at all, but I'd prefer to have some sort of basic test
that at least confirms that all options are accepted, the option values are
deposited in the correct locations, and option parse errors are handled
correctly. It might be possible to genericise such a test to be
parametrized across all commands.