src/mc68/pmon: Genericise command-specific options parsing tests
Now that `examine` and `regscmd` both use `qcmdopts`, the majority of the
testing in test_examine_parseargs() and test_regscmd_set() is redundant,
being covered by test_qcmdopts(). So replace those two with a much simpler
test that simply confirms that all known options are accepted and that
their values go into the right place.

The test also checks the parse error (abort) path, though I think that this
is not really necessary since it's not influenced at all by the caller (at
least in theory). But at the moment, it gives me a bit of comfort.