src/mc68/pmon Ccall: Restore/save registers and CCR before/after call
We save/restore only A/B/X/CCR, for reasons described in the code.

What we save and how we handle the saved PC and SP in general was not
obvious, and required some careful thought. This is all documented in a new
large comment at the head of pmon.a68. (This information may move to user
documentation at some point.)

There is still a case we're not sure how to handle: the saved PC for entry
method 2 (JSR into monitor). If we were entered with a JSR we might save
the return address that's on the stack, but how do we tell if we entered
via a JSR and not a JMP? That does not affect the call command, however.