b8tool/: Python code to assemble/link files with ASxxxx
This replaces the Bash code in `Test` that previously did this. This
is the first step towards moving the current build system into Python,
with the long-term goal of extracting the build system to a separate
repo so it can be used in multiple projects.

This also moves the ASxxxx object files from their own separate
subdirectory (.build/obj/asxxxx/$srcname/) to the subdirectories
matching the source path (.build/obj/src/$srcname.*), as the
Macroassembler AS builds do. I don't recall why I previously separated
these into their own tree or used a separate subdir for each source
file, but it was probably related to convenience in implementation.

We still, however, link with CWD set to the subdir, which is not quite
correct (see the code), but too much work to fix right now.

A `-c` build is advised when moving to this commit.