b8tool/: Replace ASxxxx build functions in Test with Python code
This also moves the ASxxxx object files from their own separate
subdirectory (.build/obj/asxxxx/$srcname) to the subdirectories matching
the source path (e.g., .build/obj/src/asxxxx), as the Macroassembler AS
builds do. I don't recall why I previously separated these into their own
tree or used a separate subdir for each source file. We still however link
with CWD set to the subdir, which is not quite correct (see the code), but
too much work to fix right now.

This is the first step towards moving the current build system into Python,
with the long-term goal of extracting the build system to a separate repo
so it can be used in multiple projects.

A `-c` build is advised when moving to this commit.