src/mc68pmon qcmdopts: Tweak non-flag option types for easier detection
Previously when checking to see if an option type is not a flag (bit) we
had to compare with each individual non-flag type. Since these always have
more than one bit set, we can make one set bit always be the MSB and then
if the option type is > $80 we know it's got multiple bits set and is not a
flag (single bit) type.