src/conftest.py: Extract loadbios() from mc68/{hello,pmon}
loadbios() now becomes a pytest fixture (and so is passed as a parameter
to a test function) that returns the loadbios() function. Doing all the
work in the returned function rather than any setup before the test starts
was a matter of programming convenience (minimal changes), but is probably
a good idea anyway so that any errors in the BIOS load/setup occur during
the test itself, rather than the test setup.

loadbios() now also takes a `biosname` parameter so that we can choose
which particular test platform (e.g., tmc68 or, one day, some 6502 version)
to use.