testmc.mc6800.opimpl SBC: Fix backwards handling of "borrow" flag
When doing an SBC, we were mistakenly adding the C flag to the result
rather than subtracting it. It's not clear if this was confusion with
the 6502 (which does things the other way around), mistyping what we
read from the _Programming Reference Manual_ p. A-59, or just not
reading that at all.

This of course changes not just the result but the flags, and the
boundary conditions become slightly different, so instead of
continuing to try to use the `SUBtests` data we instead create a new
(but similar) test data set.