testmc.mc6800: Fix broken TSX/TXS instructions.
TSX is supposed to to return the stack pointer _plus 1_ in X, and TXS is
supposed to subtract one from X before putting it in the SP. (I.e., X
points to the most recent value in the stack, whereas the SP points one
below that.)

As well as the trival fix to the simulation code and more extensive changes
to the simulator tests, various bits of user tests and code also needed to
be fixed:

- src/mc68/simple: A new test was added to help demonstrate the TSX/TXS
  behaviour. This can also easily be copied to a real 6800 system and the
  results examined.
- src/mc66/continuation: We need to change our offsets from X by one.
- src/mc68/pmon: perr_stack is now offset by 1.

Though all its tests are now passing, there are still various issues in
src/mc68/pmon: there's no test for the DEX in `init` and the registers
command is probably displaying the wrong value for the SP, for example. But
this makes little difference since missing functionality (e.g., registers
command doesn't yet set the SP itself) means that this makes no practical
difference yet anyway.