src/mc68/continuation: Test that we continue in the right place
test_Dpushcontcur() was giving a bit of coverage for this, but
test_pushcontcur(), which should have been covering this, was not because
we weren't checking the PC was correct after it was complete. You could
tell by removing one of the INX instructions used to skip the PC past the
JSR instruction and its operand.

WARNING: The TXS/TSX instructions are currently broken in the simulator so
this code won't actually work on a real 6800. This will be fixed in a
future commit.