b8tool.tool: OS-independent way of checking dependencies
We now have a command for each dependency that determines whether or not
it's present; this is usually `pkg-config libfoo` or `foo --help`. If it
fails, we print the command that failed and suggest at the end a list of
packages (one for each failure) that should be installed.

This also fixes the problem where if you had GraphicsMagick installed,
which provides what we need, the build system would still complain that
it wanted ImageMagick.

Currently these are by convention still Debian package names, though the
package names no longer matter since they're only displayed to the user.