exec/tmc68/pmon: Raise stack start slightly to avoid unused byte
Due to the previous bug in the simulator's TSX we had to set the stack one
lower than we would on a real machine. Fix that so the stack now starts
right below the readline buffer on both. The wasted byte is not so
important, but this also will make things a bit easier to understand when
debugging, particularly for counting stack usage.