Test: No `cd` except for final pytest run
b8tool does not rely on the current working directory (CWD),
using only $B8_PROJDIR, and the same is true of the remaining
build code in Test. The only thing now that relies on the CWD is
or invocation of pytest, for various things documented in the
file.

In order to show the need for $B8_PROJDIR/lib/ in PYTHONPATH this
must be tested by selecting test files not under lib/, e.g., via
`./Test src/c64.pt`. If anything under lib/ is tested pytest will
automatically add lib/ to sys.path and you will not see
src/conftest.py fail.