testmc.__init__.sym_tid(): handle non-numeric values
Since this gets called for all parameters or none, it can be very awkard to
avoid sym_tid()() being called on non-numeric values, on which to this
point it blew up. Fix that and make it reasonable to use on all parameters
by passing the value on to tmc_tid() if a symbol wasn't found for it.

Further, to avoid noise like "BIGENDIAN,BRANCHEXT,COMPMODE,FALSE,HASDSP,
HASFPU,HASPMMU,INEXTMODE,INLWORDMODE,INMAXMODE,INSRCMODE,INSUPMODE,PACKING,
RELAXED,Z80SYNTAX" (yes that really happened for 0), don't do a symbol
lookup for numbers 5 or less.