WIP reader switch to auto-ZP allocation and include objects.a65
This will all need to be cleaned up later, of course, especially because we
want to be able to fix our ZP locations for any particular platform. But
that will probably require we have the unit test builds able to specify
their own ZP equates so they no longer have to be in the assembly files
that get included by both the unit test program and other programs. (And
the unit test programs must do the includes, not the .a65 modules!)