b8tool: Major rework of asltest to build from source in the .pt file
This reorients our builds of test binaries to be based on the test module
(.pt file) rather than an assembly source file.

The binaries we build for unit testing include three types of source code:
- Code under test (CUT) that is the same as when used in production (except
  perhaps for location).
- Definitions and sometimes code expected to be supplied by the client
  program, such as the assembluy origin, locations of variables, and I/O
  routines.
- Definitions and code needed only by the tests.

Previously the latter two were supplied in the following ways:
- A few things, such as the cpu type, origin and a UNIT_TESTING=1
  definition, were generated by `b8tool asltest`.
- Some was in the same assembly file as the CUT, enabled by `ifdef
  UNIT_TESTING` blocks.
- Sometimes there was a second source file containing only test rig code
  that imported the CUT file, and this is the file that was tested.

This was awkward, and part of this was due to having the wrong orientation:
treating an assembly file as the "root source" and finding the test file
based on that.

Now we've reversed this, storing all test-related code in the test module
itself and passing that file to asltest, which starts there when building
the test binary.

The test module can now contain a `test_rig` global which contains all test
rig code and should explicitly include files with the CUT. A directory is
created under .build/ptobj/ that's parallel to the location of the .pt file
(relative to B8_PROJDIR), a new source file with the contents of `test_rig`
is generated there, and it's assembled. src/conftest.py also understands
this scheme and, if a test module being run contains a `test_rig` global,
the corresponding binary will be automatically loaded. (This has let us
delete a couple of source files that had only test rig code in them, as
well as removing all `ifdef UNIT_TESTING` blocks.)

We maintain the ability to load specific binaries as well; if an
`object_files` global (a string or sequence of strings) exists in the test
module, those files will be loaded in sequence before the test_rig file, if
any. This is useful for testing binary modules that are expected to be used
elsewhere in the test system, such as the testing BIOS. Relative paths are
relative to path.obj(), but paths with `..` or absolute paths can be used
too if necessary.

Loading (and thus executing the top level of) test modules before the code
has been assembled introduced some new issues:
- We no longer can have access to symbols at the top level because the
  source that defines the symbols hasn't yet been compiled. So we go back
  to specifying symbol names as strings when we need them at the top levelo
  (e.g., for @pytest.mark.parametrize values) and doing the lookup only
  after the test has started running.
- We do not have the additional search paths added to sys.path that pytest
  adds, so we now explicitly add B8_PROJDIR to sys.path when b8tool starts.
  (We now import shared test code as `src.stdtest` and so on, and the
  pytest.register_assert_rewrite() have been updated appropriately.) This
  may have additional implications that need to be looked at.

A `-v` verbose option was hacked in to b8tool to help with debugging.
Calling vprint() will print only if the specified number of `-v` options
have been seen.

Todo:
- `object_files` should take path-like objects as well.
- Check for test modules still using `object_files` when they would be
  better off using `test_rig`. Or perhaps some of these should be moved to
  `exe`.
- Various source files under src/ are still using `include` or doing data
  allocations; both these should be done by the client code (the test rig
  or the actual program using those routines).